requirements.txt
fastapi
uvicorn
python-dotenv
pinecone
google-genai
groq
langchain
langchain-text-splitters
pydantic
python-multipart
httpx
supabase